Data Gathering prep (Questionnaire)
Hey guys, attached is the questionnaire for the Data Gathering Prep Assignment. Please check out the questionnaire and feel free to make any comments or changes to the Questionnaire. I've already turned it it, but if you want to make changes or add questions to the questionnaire, feel free to do so and resubmit it on Canvas. Thanks guys.